requirements.txt
PyYAML==6.0
numpy==1.24.3
opencv-python==4.7.0.72